Add interface to allow for the display of study comments from Orthanc API
Provides a "Study Comment" interface which allows for users to view and add comments to a study. Comments in Sonador are used for tracking the history and storing information about resources. This merge request provides API methods and updates to the comments interface for the display and rendering of Markdown comments.